requirements.txt
ollama
ollama
pyinstaller
rapidocr_onnxruntime
Pillow
pyperclip
keyboard
win11toast
